Changes in swapping fragments and created folders
