+++ app/src/main/res/layout/fragment_wallet.xml
@@ -8,6 +8,6 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+        android:text="@string/wallet" />
 
 </FrameLayout>
