+++ app/src/main/res/menu/navigation.xml
@@ -14,7 +14,7 @@
     <item
         android:id="@+id/navigation_overview"
         android:icon="@drawable/ic_assignment_black_24dp"
-        android:title="@string/overview" />
+        android:title="@string/title_overview" />
 
     <item
         android:id="@+id/navigation_notifications"
