+++ app/src/main/res/values/strings.xml
@@ -2,16 +2,19 @@
     <string name="app_name">iExpens</string>
     <string name="title_home">Home</string>
     <string name="title_notifications">Notifications</string>
-    <string name="overview">Overview</string>
-    <string name="wallet">Wallet</string>
+    <string name="title_overview">Overview</string>
+    <string name="title_wallet">Wallet</string>
 
 
 
     <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+
     <string name="_0">0%</string>
-    <string name="title_wallet">Wallet</string>
+
     <string name="click_me">Click me</string>
     <string name="click_me_to_increment">Click me to increment</string>
     <string name="amount">Amount</string>
+    <string name="add_bills">Add Bills</string>
+    <string name="wallet">Wallet</string>
+    <string name="overview">overview</string>
 </resources>
